relocate domain files
