+++ src/features/select-pokemon-type/domain/entities.ts
@@ -1,0 +1,4 @@
+export interface IPokemonTypeItem {
+  name: string;
+  url: string;
+}
